agentic-rag-enterprise-main/requirements.txt
# Core
python-dotenv==1.0.1
pyyaml==6.0.2

# PDF Processing
pymupdf==1.24.10

# Embeddings & Vector Store
sentence-transformers==3.3.1
faiss-cpu>=1.7.4
torch>=2.0.0

# Keyword Search
rank-bm25==0.2.2

# Re-ranking
transformers>=4.40.0

# Agent Framework
langgraph==0.2.60
langchain==0.3.12
langchain-anthropic==0.3.1
langchain-openai==0.2.14
langchain-community==0.3.12

# Evaluation
ragas==0.2.8
pandas==2.2.3

# UI
streamlit==1.40.2

# Utilities
tqdm==4.67.1
numpy>=1.24.0
